python indent: do not override after parenthesis
Problem:  Vim overrides user-specified dedentation after parenthetical
          statement.
Solution: Check for user-specified dedentation, as is done elsewhere.
          (closes #6060)

Consider:
```
class Foo:
    def foo1():
        return (1
                + 2)

    def bar1():
        pass

    def foo2():
        return (
            1,
            2,
        )

    def bar2():
        pass
```

Without the patch, vim will prefer to indent the `def bar1` and `def
bar2` lines to the same level as the preceding `return` statements, even
though that is clearly not desired.

To fix this, only indent when the user has not dedented. Factor this
common check out into a function rather than repeating it.