set comments and commentstring for .asm files
Useful with tpope's commentary plugin.

There is no ftplugin for asm, so I haven't contacted the maintainer.

This might be controversial because asm has a bunch of different
dialects and `;` is only a comment in some of them. This is also true of
the current default of `/* */`, but I think `;` is right more of the
time (used in nasm, yasm, the official ARM assemblers, MASM, etc).