use attenuated ceil rounding when calculating pango char_height
This helps when the reported pango line height is just an epsilon larger
than the nearest multiple of PANGO_SCALE.
This happens frequently for bitmap fonts, whose height is a multiple of
PANGO_SCALE and will be an epsilon smaller or larger than the nearest
multiple of PANGO_SCALE.
For example I'm getting these numbers for a bitmap font of pixelsize 15:
ascent = 12293
descent = 3068
12293+3068 = 15361
15361/1024 = 15.000976562
15*1024 = 15360

this is only 1 larger then the nearest multiple and should _not_ be
ceil rounded.

Therefore this PR will floor round any value that is at most 2 larger
than the nearest multiple.