Allow completion of locales on Windows
On Unix we use the output of `:locale -a` to generate the completion
list.

On Windows, we currently have only expansion of the arguments `mess time
and ctype` for the :lang command.

This patch examines the directory structure of $VIMRUNTIME/lang/ and
uses that as valid locales (and in addition it adds the value of C). As
a side effect, we can then also enable the :lang completion test for
Windows.

There is one specialty however, do not add locales that contain the `.`
which is used delimit the encoding from the locale. So if we have directories 'ja',
'ja.euc-jp' and 'ja.sjis' only allow completion for 'ja'. I am not sure
this is correct however(?)