Arglist: quit when all args in local arglist have been seen
Issue #6133

When using a local arglist and we have seen all args in the local
argument list, Vim will still ask for confirmation, because the global
arg_had_last flag was not correctly set.

So before creating a new window-local argument list, check if there is
actually a global argument list, and set the arg_had_last flag if there
isn't or we are currently editing the last file in that list.