Support left/right mouse wheel when running in a terminal
This allows you to scroll sideways with the left/right mouse wheel when
nowrap is set, and to map <ScrollWheelLeft> and <ScrollWheelRight>. This
(obviously) only works in terminal emulators that support it (like xterm
and urxvt).

Note that scrolling sideways only work when vim is compiled with gui
enabled (apparently that affects the terminal version too), but mapping
the keys works either way.