win32: Fix window positioning
This patch was originally written by @ynkdir:
https://groups.google.com/d/msg/vim_dev/xRdOvRDM3-Y/Ej-vJ0dWzPoJ
and listed in the todo.txt:
```
MS-Windows resizing problems:
- Windows window on screen positioning: Patch by Yukihiro Nakadaira, 2012 Jun
  20.  Uses getWindowRect() instead of GetWindowPlacement()
- Win32: When the taskbar is at the top of the screen creating the tabbar
  causes the window to move unnecessarily. (William E. Skeith III, 2012 Jan
  12) Patch: 2012 Jan 13  Needs more work (2012 Feb 2)
```
I updated it so that it can be applied to the latest code.

Window positioning was wrong when the taskbar is placed at the top or
the left of the screen. This is because Get/SetWindowPlacement APIs are
using different coordinates from the other APIs. When the taskbar is
placed at the top or the left, the origin for Get/SetWindowPlacement is
shifted to the bottom or the right of the taskbar.
Using GetWindowRect and MoveWindow APIs fixes the problem.

I also confirmed that this works on a multi monitor environment.