win32: Support snapping on the edges of the screen
Starting from Windows 7, MS-Windows supports Aero Snap (and it is now
called just Snap). It can snap windows on the edges of the screen.

If a gVim window is snapped, and then the window size is changed by a
command like `:vsp` or `:tabe`, snapping will be cancelled. This might
be annoying for some people.
Actually this can be avoided by setting `:set go+=k`, but this also
changes the behavior when the window is not snapped.

This patch detects if the gVim window is snapped, and acts like if the
window is maximized. So, if the window is snapped, and even if a command
like `:vsp` is executed, snapping will be kept. But if the window size
is explicitly specified (e.g. `:set lines=30`), the size will be changed.

Related: #123, #703, #2180