Fix visual block gap on scroll with scrolloff
After scrolling down using CTRL+D, one win_update() is called and the
following steps will happen in order:

  1) Propagate the line informations that is not above the wp->w_topline
     entries upwards.

  2) Calculate the range of lines affected by the scroll.

  3) Updates wl_lnum of affected lines.

In some cases, win_update() starts with buf->b_mod_set set to true. One
example where this can happen is when two scroll happens, the first
scroll results in the cursor pointing to a parentheses/bracket and the
second scroll results in the matched parentheses/bracket positioned
right at w->w_topline. If matchparen autocommand group is active, it
will highlight and clear the parentheses on first and second scroll
respectively. This will set buf->b_mod_set to true.

The first scroll generates a CursorMoved event that gets applied by
matchparen autocommand, which will result in a call to match_add() and
set buf->b_mod_set to true. The second scroll generates the same event,
in this calling match_delete() and set buf->b_mod_set to true.

When buf->b_mod_set is set to true and the modified line is positioned
right at w->w_topline, the step 1) of win_update() does not happen
because the first conditional will be be chosen instead. This results in
stale wl_lnum for lines above wp->w_topline.

The calculation done on the second step requires the wp->w_topline to be
found. However, if the propagation step does not happen, the line that
correponds to wp->w_topline will be situated are the line with higher
index than the reality, hence the resulting rendering gap. This doesn
not happen when scrolloff is not set because the lines are drawn until
the end anyway.

This commit fixes this by making sure we propagate the line if the
window has definitely scrolled.