testing: Always show messages
"messages" was shown only if it contained "SKIPPED" or "FAILED".
However, now "messages" contains information about execution time.
It is useful always to show "messages".

Also there was an issue in Make_dos.mak and Make_ming.mak.
If either "SKIPPED" or "FILED" was not found in "messages", `findstr`
returned 1, then the test was failed.