Fix macOS default locale to have UTF-8 encoding set
If LANG is unset in macOS (either manually or when launching GUI version
of Vim from the Dock where this environmental variable is not set by the
OS), Vim tries to query the system locale and set that. However, the
system API doesn't set the encoding part, and there are tools like new
versions of Python that seems to want to have an explicit encoding bit
set. Also, macOS defaults to UTF-8 anyway, so just make sure to append
".UTF-8" to the locale to explicitly set it.

Also, add regression test for both this change and for 509f8031 (#7003)
which added logic to make sure we set LC_NUMERIC to "C" in the default
locale detection.

This is a port of https://github.com/macvim-dev/macvim/pull/1036