win32,python3: Fix python3 stdin issue
This fixes the python3 issues related to the following issues:
 #6641, #7053, #7061, #7083, https://github.com/vim/vim-win32-installer/issues/193

In the first issue #6641, I reported that vim crashed when stdin was
redirected, but it turned out that it was not correct. Vim crashes when
it reads the input from stdin when the `-` option is used.
When vim is executed by `vim -`, vim closes stdin in fileio.c and main.c
by `close(0)`. Python 3 cannot use stdin because of this, and that's why
python 3 aborts.

This PR checks if stdin is readable, and if not, reconnect stdin to
"CONIN$".

Note that the following patch is not needed anymore:
https://github.com/vim/vim/commit/253b16a4abdad1df350b9ddd9a709520b063934c#commitcomment-43022526