Single quotes, not triple double quotes in :h undo
In `:h undo` most of the examples use single quotes when referring to
registers, e.g. '"2P', since this is probably more readable/clear.

The first example however, uses ""1P", which can make the second double
quote (the one that actually needs to be typed) harder to spot.

This simple change will make it easier to readm, and more consistent
with the other examples.