make ceil-rounding of the char-box height less aggressive
Turns out, the 3‰ tolerance for ceil-rounding is too little.
I did some calculations about possible rounding errors. Since OpenType
fonts use a typical em-size of 2048 and pango uses 1024 units per pixel,
the rounding error in each of ascent and descent may be up to
half the size of the font in pixels, adding up errors in ascent and
descent, we need to expect rounding errors up to the font size in
pixels.
I would suggest increasing the tolerance margin to 1/16th pixel
seems to be a sensible trade-off to me.