if_perl: Fix that Perl 5.32 doesn't work when compiled with MSVC
When if_perl is compiled with Perl 5.32 and MSVC x64, the `:perl`
command shows the following error:
```
auto/if_perl.c: loadable library and perl binaries are mismatched (got handshake key 0000000012800080, needed 0000000012900080)
```

This is because the size of the `PerlInterpreter` structure is different
when compiled by MSVC and MinGW.
This fixes it by defining `NO_THREAD_SAFE_LOCALE`.

Also adjust some comments.