Require the swap file to be from the same computer to auto delete
Resolve part of a TODO, I'm arguing yes, you can't check that the
process is still running on another computer.  For NFS auto deleting
the swap file removes the reminder that you might be editing the same
file on another computer, and you can no longer find out what computer
you were editing it on without disabling auto delete.

I don't have an argument for checking the username.  It is now
checking the hostname, and that it is not running, saying someone else
is interested in the file might be useful, but their editor isn't
running and there aren't any changes.