Removed legacy dependency on win32.mak
In the prior version of this makefile there was a reference to win32.mak.   I went down a rabbit hole of trying to find this file because I was led to believe (from other sources) that it was important.  This file - win32.mak - is not part of the last 2 major releases of Windows SDK and they are no longer making a downloadable version of Windows 7 SDK so finding it will be non-trivial for most folks.

As it turns out from the comment by k-takata in
  https://github.com/vim/vim/issues/3023

none of the useful variables defined in win32.mak are actually used in make_mvc.mak anymore so for general code hygiene I've removed it.   I've confirmed that it still builds on my setup here and am taking k-takata's comment as the truth for all of the other build variants that might currently be supported.