Make window popup flag logic conditional
Patch 8.2.1990 (commit 6a07644) adds some flag switching on the w_flags
member of the current window, but this member is only part of the
relevant struct if Vim is being compiled with +popupwin, and so
compilation fails if the feature is excluded.  This commit adds
a compile-time condition around the new code so that Vim can be built
without the feature.