Revise test_swap
Test_swap_auto_delete() still often fails on MS-Windows.
This executes 'echo' as a temporary job, and uses its pid as an unused pid.
This assumes that a pid which is used once will not be reused soon.
If it fails to execute a job, just add 4 to the current pid.