Don't trigger TextYankPost on X external selections
Due to how Xorg clipboards work `clip_get_selection` is be called
whenever another client requests the clipboard content owned by Vim.
Prevent the emission of autocmds when the yank operation is triggered by
some external mechanism, this avoids nasty side-effects such as #4201
beside stopping the emission of spurious TextYankPost events.

A test case to make sure a TextYankPost is triggered when the visual
selection is left (and cb=autoselect or cb=autoselectplus) is also added
to prevent this feature from regressing.

Closes #4201