GTK3: fix white space around text in another way
In the old way setting background color for the whole window affects
menubar background.

The white space belongs to the GtkForm containing the drawarea, so let's
set background color on that. The GtkForm also contains those scrollbars
but they have their own background color and thus not affected. (Whoever
changes part of the scrollbars to transparent should have considered the
background color of widget below since scrollbars can appear on vast
kinds of widgets.)

This fixes #7427.