Add arch (x86_64 / arm64) to :version output for macOS
With the introduction of Apple Silicon and Rosetta (x86 translation on
ARM), there are now two common architecture for Vim on macOS and it's
frequently useful to see which arch Vim is running under to see if it's
under Rosetta or native ARM. Add architecture info for macOS version
output.

Not sure if we should do something like this for Windows eventually as
well as ARM builds may be more common.