Mention scroll option has been reset by changed window size
Slightly change the previous logic. Instead of resetting Script Context,
remember it has been reset by a changed windows size and therefore
mention this in the `:verbose :set scroll` option.

In addition, fix a tiny built error, by ifdef'ing those changes.

Adjust the test and in scriptfile mention what the Output string is, so
it is easier to add new messages.