regex: do not bail out if regex_engine=auto for char classes
The documentation for [[:upper:]] states, that those only work for 8 bit
characters for the backtracking regexp engine. So for character classes
such as [[:upper:]]\+, the NFA regexp engine should be used.

Unfortunately, when using the auto regexp engine, compiling the regexp
fails, because there are too many states. Vim will then fall back to the
BT regexp engine, which unfortunately cannot handle character classes
for multibyte characters.

So when compiling the regexp, set a flag `wants_nfa` and when testing
whether there are too many states for the NFA engine, do not bail out,
if the wants_nfa flag has been set.

In addition, the check for a too large maxval should also check whether
maxval equals MAX_LIMIT, because in that case the NFA engine emits an
NFA_STAR or NFA_STAR_NONGREEDY atom, and doesn't use too many states. So
slightly adjust the condition for bailing out for those values.

Add a test to document, that both engines return different number of
matches in such a case.