regex: swap condition for writing debug log for BT engine
When writing the debug log for the BT engine, the test currently tests
for the regexp engine being different of the backtracking engine, while
the code around it (the ifdef and the logfile name) actually assume the
logfile for the backtracking engine is written.

So invert the condition and mention in the comment that this is writing
the log file for the BT engine.