fix asan error when closing window causes buffer-overflow
As can be seen here:
https://github.com/vim/vim/runs/1623161184?check_suite_focus=true

ASAN still reports an error, when closing a window, which causes a
destroy scrollbar event, which in turn will cause a resize event which
will finally try to set the shell window and might access already freed
variables.

So fix this by checking, whether the current window is still valid. If
the current window is in the process of being freed, wp->w_lines will be
set to null by win_free before handling the scrollbar event.

That should fix the ASAN error hopefully.