Complete *func and *expr options
All options that take a function as their value:

	completefunc
	imactivatefunc
	imstatusfunc
	omnifunc
	operatorfunc
	quickfixtextfunc
	tagfunc

Now complete with with EXPAND_USER_FUNC, which excludes built-in
functions as this probably makes the most sense.

And all options that take an expression:

	balloonexpr
	charconvert
	diffexpr
	foldexpr
	foldtext
	formatexpr
	includeexpr
	indentexpr
	patchexpr
	printexpr

Complete with the new EXPAND_EXPRESSION_SET; this is mostly identical to
EXPAND_USER_FUNC except that it adds "()" at the end.

There is already EXPAND_EXPRESSION, but in my testing I didn't really
find that too useful since it can't deal with :set's syntax; neither of
these work for completing F2():

	:set balloonexpr=F1()\ ..\ F2()
	:set balloonexpr=F1()..F2()

It only completes the F2 when you add spaces:

	:set balloonexpr=F1() .. F2()

But that's not valid syntax for :set.

This could be expanded later on to deal with all of this, but IMO this
is already an improvement over not having any completion at all.