Add findcommand() function
Right now there is no way to know which shortcut commands refer to which
commands. This adds findcommand() to solve that.

Context: I made a little VimScript to show the help for the command you
typed: pressing <F1> while you typed ":syntax" shows the help for
:syntax. But I'd also like this to work for all the shortcut commands,
which is hard without this patch, especially if you want it to be a bit
smarter (i.e. load ":syn-match" if you typed :sy match").