Improve accuracy of the normal % command when buffer has many lines
The normal `{count}%` command uses the following formula to
jump to a percentage of lines with rounding up (so 100% jumps
to the last line):
```
  ({count} * number-of-lines + 99) / 100
```

The above formula can cause 32-bits overflows for large number
of lines, so vim uses another formula, with degraded accuracy,
when the buffer has > 1,000,000 lines:
```
(number-of-lines + 99) / 100 * {count}
```

As a result, the following command jumps to line 500050 instead
of the more accurate rounded up line 500001:
```
$ vim --clean -c'call setline(1,range(1,1000001))' -c'normal 50%'
```
Yet 1,000,001 lines would not cause a 32-bits. Overflow would
happen starting with the much larger 21,474,836 lines threshold.
So this PR changes `nv_percent()` to use a larger threshold beyond
which to use the formula with degraded accuracy.

After this change, the above command jumps to line 500001
as expected rather than 500050.

I did not add a test with 21,474,836 lines because it would
consume too much memory and would be slow.