os_unix.c: Use sysconf() to get SIGSTKSZ
The newest glibc - 2.34 - sets SIGSTKSZ constant to -1 when _GNU_SOURCE
or _SC_SIGSTKSZ_SOURCE are defined. This change causes Vim to segfault,
because it now treats the negative -1 as a positive number, which is the
highest number available.

The signal stack size is now available via 'sysconf(_SC_SIGSTKSZ)',
which is used in the commit, but we need to extra check if this way is
available in configure phase. The local SIGSTKSZ #define will be used
only if SIGSTKSZ isn't already defined and if it is not available
through 'sysconf()' at the same time.