doautoall: Execute autocommands for the current buffer at last
The doautoall command invokes an event on all the buffers, but some event
listener may store the current window. For example, doautoall WinEnter confuses
some plugin detecting the current window for changing the statusline.
The document of doautoall command does not mention the order of the buffers so
moving the invocation on the current buffer at last will help various plugins.