Prevent spell checking within nroff constructs
Almost all nroff constructs are non-spelling-checkable abbreviations,
so like html.vim we use toplevel/@Spell-on-one-item to prevent spell
checking in all syntactically-significant text. As for several other
formats, comments are a natural choice for reactivating spelling on.
Also reactivate within .TS/.TE and .[/.], as tbl and refer contents
are primarily text.

Also be explicit that these syntax regexps are case sensitive.