Call breakcheck less often when adding completions
There are many different ways to add completions to the popup menu, some
of which are slow (such as scanning the filesystem). However, a number
are very fast loops, such as the vimscript complete() function and
various ways to add completions from a predefined list or dict.

In the fast cases, the calls to breakcheck can start to dominate
performance as it can be quite slow to call select() on some platforms.

Solution: Call it less often in these cases.