fix interpretation of `TOP` in `syn include`d file
When running `syn include @NAME FILE`, the `TOP` in the sourced `FILE`
should be interpreted as `@NAME`, not the `TOP` of the caller.

For example, consider markdown fenced code block containing C code. To
highlight the C code in the block, we would write something like this in
syntax/markdown.vim:

    syn include @FENCED_C syntax/c.vim
    syn region start="```c" end="```" contains=@FENCED_C

In this case, this line in $VIMRUNTIME/syntax/c.vim

    syn region cCppOutElse ... contains=TOP,...

should be should be interpreted as

    syn region cCppOutElse ... contains=@FENCED_C,...

because we want `cCppOutElse` to contain the `TOP` of C syntax, not the
`TOP` of markdown syntax.