Use POSIX style redirect for ash and dash in shellpipe
Update comment to say "POSIX style redirection" since the `2>&1`
redirection is specified in POSIX and is not a bourne shell specific
thing[1][2].

Add busybox ash and dash to the list of known POSIX compatible shells so
stderr is captured when SHELL is set to one of those.

Add dash.exe to windows but not ash, since busybox ash does not run on
windows.

[1]: https://pubs.opengroup.org/onlinepubs/9699919799/utilities/V3_chap02.html#tag_18_07_02
[2]: https://pubs.opengroup.org/onlinepubs/9699919799/utilities/V3_chap02.html#tag_18_07_06