Encryption: Provide alternative using libsodium
After the issues #638 and #639 I thought about exploring alternatives to
make the crypt code more modern and robust. So for the last couple of
weeks and spend my spare time exploring things and improving my crypt
skills ;)

One of the mentioned alternatives was
[libsodium](https://libsodium.gitbook.io/doc/), which is a
library providing encryption, hashing and other cryptographi features
for end-users. Fortunately it is cross plattform so can be used across a
variety of platforms.

This is currently work in progress, it compiles and seems to work for
me, but is not yet considered stable. Also, I am currently not sure, the
changes to memline/memfile are actually correct. I don't really that
low-level code of the data blocks :/

The reason I am posting it here is, that I'd like to get feedback about
the current scope.

In addition I am not so sure, how to properly hook into encrypting the
blocks for swapfiles and undo-files. For stream-encoding, libsodium
needs to know, when a stream of messages is completed and adds an
additional tag for verification. However given the current design of the
block-writing code, I am not sure how to detect whether the current call
for encrypting will be the last one. So I would like to have some
guidance here.

For now, I have just sprinkled a couple of static TRUE and FALSE for the
swap file encryption and for undo-file encryption I just disabled
complete encryption.

Also the current read-file and write-file end up using different block
sizes, so I made some slight changes to the read-file code, to ensure,
that blocks of 8K will be read to make sure that the crypt layer can read
the same block size when reading as when writing. Not sure if there is a
better way to achieve this.

And of course, I'd like to get some CI builds done :)

I'll have a look at the windows changes, once I get my windows build
system ready. Not sure how much work this requires.

Tests are written and enabled in the CI system.

Also, a couple of minor bugs have been fixed in the reading and writing
of buffers code.

Todo:

- [ ] fix error numbers
- [ ] write tests
- [ ] make changes to the windows build system
- [ ] enable complete undofile encryption (?)