When the defaults.vim file cannot be read, output an error message
When running vim --clean and the `defaults.vim` runtime file cannot be
loaded, output an error message instead of silent to continue.

That is because if defaults.vim cannot be loaded, vim will start in
defaults mode, which may be unexpected and behave very different. So
make the user aware of this situation.