fix syntax highlighting of vim option
It might be broken when on a line with an escaped trailing character.
In that case, the end of the region cannot be found because of the skip
and oneline arguments.  Fix the issue by tightening the skip regex; do
not skip an escaped character at the end of a line.