Define and use ARRAY_LENGTH macro
The "(sizeof(arr) / sizeof((arr)[0]))" idiom was used very often in the
Vim source code.  Also ARRAY_SIZE, NUM_ELEMENTS and some other macros
were defined.  Unify them into ARRAY_LENGTH and use it everywhere when
we need the size of an array.

For MS-Windows, MS defines the ARRAYSIZE macro, so we can use it in the
Windows-specific files.