add syntax for Graphviz’ GVPR format
Graphviz, the graph visualization software, comes with its own scripting
language for graph manipulation, “GVPR.” The syntax support for this language
that is added in this commit comes from the Graphviz repository. It was written
and then refined within the Graphviz repository, intended to be eventually
proposed for inclusion within Vim itself.