filetype: setf json for *.json-patch (RFC 6902)
JSON Patch defines a JSON document structure for expressing a sequence
of operations to apply to a JavaScript Object Notation (JSON) document.
It is specified by [RFC 6902], which specifies `.json-patch` as the file
extension for JSON Patch files in Section 6.  It seems unlikely that the
file extension would contain anything other than JSON.

Therefore, use the json filetype for *.json-patch files.

[RFC 6902]: https://datatracker.ietf.org/doc/html/rfc6902

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>