update xxd comments and version string
The changelog in the top comment record multiple changes between 1998 and 2020. Thanks to everybody for contributing!
I'd like to change some wording in the comment and in the version string:
- update my email-address,
- strip "Small changes made afterwards" -- I want to avoid to diminuish any contributers achivements.
- remove the silly poem, which atributed credit exclusively to myself.

Updates to the version string were neglected in the past. I agree with Bram, that the version number 1.10 is now meaningless. 
I propose to put the date in the form yyyy-mm-dd of the last change instead of a version number.