crypt: reset undofile option when writing files without encrypting undofile
This is left over from the new sodium feature (see
https://github.com/vim/vim/issues/8467#issuecomment-872971450)

So let's reset the undofile option when actually writing the buffer and
delete an existing undo file in case it exists. This is so, that no
information is leaked from the undo file.

While looking at it, slightly adjust the logic from commit
65aee0b714e809b9f19862f3decd35055ed4de10 (v8.2.3063) and use the buf
pointer instead of the curbuf).

This fixes #8467