Add bufnr to prop_list() and prop_find()
Getting properties with prop_list() don't return a buffer number:

	:echo prop_list(line('.'))[0]
	{'id': 0, 'col': 2, 'end': 1, 'type': 'keyword', 'length': 5, 'start': 1}

Getting this property fails as it's defined for this buffer:

	:echo prop_type_get('keyword')
	{}

So I need to do:

	:echo prop_type_get('keyword', #{bufnr: bufnr('')})
	{'highlight': 'Keyword', 'end_incl': 0, 'start_incl': 0, 'priority': 0, 'bufnr': 1, 'combine': 1}

But it's not clear from the prop_list() return value that this is the
case: I need to try both.

The same issue exists with prop_find():

	:echo prop_find(#{type: 'keyword'})
	{'lnum': 10, 'id': 0, 'col': 2, 'end': 1, 'type': 'keyword', 'length': 5, 'start': 1}

This fixes the issue by adding the type_bufnr to the return value of
both functions. I used type_bufnr because I thought that was a lot
clearer than just bufnr, as "bufnr" can refer to either the bufnr of the
property or of the property type (adding the bufnr of the property
itself in the return value doesn't make much sense, but it's easy to get
the two mixed up).

There is still a bit of an issue with this as this will fail:

        # E158: Invalid buffer name: 0
        var props = prop_list(line('.'))
        echo props->mapnew((_, v) => prop_type_get(v.type, {bufnr: v.type_bufnr}))

You need to do:

        echo props->mapnew((_, v) => prop_type_get(v.type, (v.type_bufnr ? {bufnr: v.type_bufnr} : {})))

Not sure what the best fix for that would be; on one hand you *want* to
get an error if you pass an invalid bufnr, but on the other hand being
able to get the prop types from the prop_list() output without a lot of
if/else checks. Maybe just leaving it as-is is fine.