insexpand.c: fix match finding stop condition
Also fix "may be used uninitialized" compiler warning. Previously
the condition was poorly written in such way that it was always
triggered when the line position matched the line position of the
first match, not when the combined position of the match was less
or equal to the position of the first match.