if_lua: stop on errors in vim.command() and vim.call()
This stops execution if vim.command() or vim.call() failed. This fixes
two problems:

The first is the lack of line numbers if something failed:

	Error detected while processing command line..function <SNR>48_Cmd:
	line   13:
	E1174: String required for argument 2
	E1174: String required for argument 2
	E1174: String required for argument 2
	E1174: String required for argument 2
	[..repeated many times..]

It's not at all clear *what* command failed where, and vim.command()
always returns 0. With this, it's:

	Error detected while processing command line..function <SNR>49_Cmd:
	line   13:
	E1174: String required for argument 2
	lpeg.vim/lua/lpeg.lua:131: lua: function failed

Which is much more useful.

The second is that if you have some error in a loop you'll be spammed with
errors; this is #8649

---

The big downside with this patch is that it changes the behaviour; I don't know
if any scripts relied on execution continuing after errors, but they might.

However, Lua itself stops execution out if a (Lua) function failed, vim.eval()
did as well, and it generally just makes sense IMO. It's really confusing and
surprising otherwise

An alternative might be actually returning something from luaV_command();
luaV_call() already does so. You'll have to add "if not vim.command(...) then
error("command failed") end" to everything, which doesn't seem very "Lua-like".

Another point of improvement is that the errors for "vim.fn" are
confusing; for the same as above it prints:

	Error detected while processing command line..function <SNR>48_Cmd:
	line   13:
	E1174: String required for argument 2
	[string "vim.fn = setmetatable({}, {..."]:4: lua: call_vim_function failed

Which isn't very good. Need to look in to that.

(I'll update the tests and docs later once the behaviour is decided)