gj may not jump to next screenline correctly
This may happen, if an extra wide character appears and the soft
line break. In that case, the cursor will be positioned at the beginning
of the several char wide character so the cursor may become stuck on the current screen
line.

Fix this by advancing to the next line if that happens.

Add a test