Added syntax for the structurizr DSL
structurizr DSL is used in structurizr, which is a tool that allows to
model your software architecture according to the c4 model
(https://c4model.com/)

The default file extension is .dsl but it clashes with an existing (but
non-compatible) filetype. I have no idea to detect the appropriate
filetype based on a file, so for now you have to use `set filetype`
appropriately.

The language reference can be found here:
https://github.com/structurizr/dsl