Escape backslash in fish shell with shellescape()
Fish shell uses backslash as an escape character in single quotes, so
the backslash must be escaped in order to produce a literal backslash.
This change modifies the `shellescape()` function to escape backslash in
Fish shell only.

Note that Fish's backslash behavior is a bit odd in that it does what it
thinks you mean. Only `'` really needs to be escaped (`echo '\''` prints
`'`), so if the backslash precedes other characters, it will not be
treated as an escape character (`echo '\a'` prints `\a`). However, all
backslashes within single quotes can safely be escaped because two
consecutive backslashes always produce a single literal backslash
(`echo '\\a'` also prints `\a`).