Properly escape character after backslash in fish
I recently added a fix to make `shellescape()` properly escape
backslashes for fish shell (see 6e82351130d, patch 8.2.3385). However,
that fix added a bug where the character following the backslash was
always added to the string as-is, even if it was a special character
that should have been escaped. This change fixes that bug.