Trigger TextChanged after TextChangedI correctly
Use a separate variable for deciding whether or not to trigger
TextChangedI autocommand or not. Else we may skip an existing
TextChanged autocommand, just because TextChangedI triggered right
before.

fixes #8932

Note the test: For some reason, evaluating the g:autocmd_n variable was
never performed. Not sure why, it works when running the test
interactively :/