Fix config of old systems for nanosecond-resolution timestamps
On old systems (e.g. 32-bit systems running Ubuntu 10.4), defining
_XOPEN_SOURCE causes _BSD_SOURCE and _SVID_SOURCE not to be defined,
which causes nanosecond-resolution timestamps not to be included in
struct stat, which causes the build of fileio.c to fail.

_XOPEN_SOURCE is defined for some systems in vim.h.

A solution is to define _BSD_SOURCE and _SVID_SOURCE whenever
_XOPEN_SOURCE is defined.

On new systems, defining either _BSD_SOURCE or _SVID_SOURCE causes
/usr/include/features.h to issue a warning message about _BSD_SOURCE and
_SVID_SOURCE being deprecated and to use _DEFAULT_SOURCE instead.

A solution for that is to also define _DEFAULT_SOURCE when _BSD_SOURCE
and _SVID_SOURCE are defined.